Move most cli tests into test_fetch; add new tests
