+++ experimenter/manifesttool/hgmo_api.py
@@ -1,4 +1,3 @@
-import os
 from pathlib import Path
 from typing import Any
